Refactor so that plugins can always add their own URLs at runtime.
This commit refactors the `get_syndication_links()` function in order to
ensure that the `syn_add_links` filter is always processed. Prior to
this commit, the filter was only processed if the `syndication_urls`
post meta field was non-empty, which meant that plugins were unable to
modify the `$urls` array that contained POSSE syndication addresses
unless a prior syndication URL already existed. But this constraint made
the plugin mostly useless for auto-crossposting plugins that wanted to
automate the process of adding syndication links to crossposted entries.

This commit also cleans up mixed spaces and tabs. :\