Fix bug where erasing all content from metabox didn't delete meta field.
This commit fixes a few related things. First, it fixes a bug where
deleting the entire contents of "Syndication Links" meta field did not
also remove the meta field from the post itself, which resulted in the
appearance of being "ignored." That is, the user woul try to delete the
content of the meta field, but it would always return. This issue is
addressed by checking for a submitted empty string and deleting the meta
field in that case.

Another issue fixed is a PHP Warning emitted that was caused by
occasionally treating the input to `syn_clean_urls` as a string, and
occasionally as an array. This issue was addressed by refactoring
`syn_clean_urls` to always expect an array as input, and moving the
specific URL validation checking to its own `syn_clean_url` function,
which expects a string.