Added Settings Information
Took a stab at delineating some of the settings for the plugin as well as providing pictures of what some of the settings do and how they'll display on a page.  

Please doublecheck the functionality description of the rel-me portions before committing. Particularly in relation to https://github.com/dshanske/syndication-links/issues/13