Create bookwyrm.svg
Info of this svg file:

- LICENSE: GPL License
- AUTHOR: Donnnno

---

BookWyrm is a social network designed for tracking your reading, discussing books, writing reviews, and discovering what to read next. It is decentralized, friendly, and anti-corporate.

The platform is federated, meaning it allows BookWyrm users to join small, trusted communities that can connect with one another, and with other ActivityPub services like Mastodon and Pleroma.